Speed up preloaded Docker test image creation
