+++ app/src/main/java/be/tomcools/advancedtestingpatterns/app/velo/VeloRepository.java
@@ -1,7 +1,5 @@
 package be.tomcools.advancedtestingpatterns.app.velo;
 
-import org.springframework.data.jpa.repository.Modifying;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
